+++ app/src/org/commcare/android/framework/CommCareActivity.java
@@ -23,6 +23,7 @@ import org.odk.collect.android.views.media.MediaEntity;
 import org.odk.collect.android.views.media.MediaState;
 
 import android.annotation.TargetApi;
+import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -31,8 +32,12 @@ import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
+import android.util.DisplayMetrics;
 import android.util.Pair;
+import android.view.GestureDetector;
+import android.view.GestureDetector.OnGestureListener;
 import android.view.MenuItem;
+import android.view.MotionEvent;
 import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -45,7 +50,7 @@ import android.widget.Toast;
  *
  */
 public abstract class CommCareActivity<R> extends FragmentActivity implements CommCareTaskConnector<R>, 
-    AudioController, DialogController {
+    AudioController, DialogController, OnGestureListener {
     
     protected final static int DIALOG_PROGRESS = 32;
     protected final static String DIALOG_TEXT = "cca_dialog_text";
@@ -62,6 +67,8 @@ public abstract class CommCareActivity<R> extends FragmentActivity implements Co
     //fields for implementing task transitions for CommCareTaskConnector
     boolean inTaskTransition;
     boolean shouldDismissDialog = true;
+    
+    private GestureDetector mGestureDetector;
     
     /*
      * (non-Javadoc)
@@ -104,6 +111,10 @@ public abstract class CommCareActivity<R> extends FragmentActivity implements Co
                 bar = new BreadcrumbBarFragment();
                 fm.beginTransaction().add(bar, "breadcrumbs").commit();
             }
+        }
+        
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+            mGestureDetector = new GestureDetector(this, this);
         }
     }
     
@@ -789,4 +800,133 @@ public abstract class CommCareActivity<R> extends FragmentActivity implements Co
         return true;
     }
     
+    /*
+     * (non-Javadoc)
+     * @see android.app.Activity#dispatchTouchEvent(android.view.MotionEvent)
+     */
+    @Override
+    public boolean dispatchTouchEvent(MotionEvent mv) {
+        boolean handled = mGestureDetector == null ? false : mGestureDetector.onTouchEvent(mv);
+        if (!handled) {
+            return super.dispatchTouchEvent(mv);
+        }
+        return handled;
+    }
+    
+    /*
+     * (non-Javadoc)
+     * @see android.view.GestureDetector.OnGestureListener#onDown(android.view.MotionEvent)
+     */
+    @Override
+    public boolean onDown(MotionEvent arg0) {
+        return false;
+    }
+    
+    /*
+     * (non-Javadoc)
+     * @see android.view.GestureDetector.OnGestureListener#onFling(android.view.MotionEvent, android.view.MotionEvent, float, float)
+     */
+    @Override
+    public boolean onFling(MotionEvent e1, MotionEvent e2, float velocityX, float velocityY) {
+        if (isHorizontalSwipe(this, e1, e2)) {
+            if (velocityX <= 0) {
+                return onForwardSwipe();
+            }
+            return onBackwardSwipe();
+        }
+                
+        return false;
+    }
+    
+    /**
+     * Action to take when user swipes forward during activity.
+     * @return Whether or not the swipe was handled
+     */
+    protected boolean onForwardSwipe() {
+        return false;
+    }
+    
+    /**
+     * Action to take when user swipes backward during activity.
+     * @return Whether or not the swipe was handled
+     */
+    protected boolean onBackwardSwipe() {
+        return false;
+    }
+
+    /*
+     * (non-Javadoc)
+     * @see android.view.GestureDetector.OnGestureListener#onLongPress(android.view.MotionEvent)
+     */
+    @Override
+    public void onLongPress(MotionEvent arg0) {
+        // ignore
+    }
+
+    /*
+     * (non-Javadoc)
+     * @see android.view.GestureDetector.OnGestureListener#onScroll(android.view.MotionEvent, android.view.MotionEvent, float, float)
+     */
+    @Override
+    public boolean onScroll(MotionEvent arg0, MotionEvent arg1, float arg2, float arg3) {
+        return false;
+    }
+
+    /*
+     * (non-Javadoc)
+     * @see android.view.GestureDetector.OnGestureListener#onShowPress(android.view.MotionEvent)
+     */
+    @Override
+    public void onShowPress(MotionEvent arg0) {
+        // ignore
+    }
+
+    /*
+     * (non-Javadoc)
+     * @see android.view.GestureDetector.OnGestureListener#onSingleTapUp(android.view.MotionEvent)
+     */
+    @Override
+    public boolean onSingleTapUp(MotionEvent arg0) {
+        return false;
+    }
+
+    /**
+     * Decide if two given MotionEvents represent a swipe.
+     * @param activity
+     * @param e1 First MotionEvent
+     * @param e2 Second MotionEvent
+     * @return True iff the movement is a definitive horizontal swipe.
+     */
+    public static boolean isHorizontalSwipe(Activity activity, MotionEvent e1, MotionEvent e2) {
+        DisplayMetrics dm = new DisplayMetrics();
+        activity.getWindowManager().getDefaultDisplay().getMetrics(dm);
+
+        //screen width and height in inches.
+        double sw = dm.xdpi * dm.widthPixels;
+        double sh = dm.ydpi * dm.heightPixels;
+        
+        //relative metrics for what constitutes a swipe (to adjust per screen size)
+        double swipeX = 0.25;
+        double swipeY = 0.25;
+        
+        //details of the motion itself
+        float xMov = Math.abs(e1.getX() - e2.getX());
+        float yMov = Math.abs(e1.getY() - e2.getY());
+        
+        double angleOfMotion = ((Math.atan(yMov / xMov) / Math.PI) * 180);
+        
+        //large screen (tablet style 
+        if( sw > 5 || sh > 5) {
+            swipeX = 0.5;
+        }
+        
+
+        // for all screens a swipe is left/right of at least .25" and at an angle of no more than 30
+        //degrees
+        int xPixelLimit = (int) (dm.xdpi * .25);
+        //int yPixelLimit = (int) (dm.ydpi * .25);
+
+        return xMov > xPixelLimit && angleOfMotion < 30;
+    }
+
 }
